Fixes JavaScript syntax highlight - Fixes #1
JSFiddle disables syntax highlight when the Javascript 1.7 language is chosen. If we choose babel it will try to compile it itself though, without the presets we want to use.

I've just found a loophole using a different between the babel code JSFiddle is using and the one we're using. Our code will now change the script type to be `text/babel;harmony=true` instead of just `text/babel`, since that's not supported by JSFiddle's transpiler, but is supported by ours. That way we can guarantee that the compilation is done with our presets, while keeping the syntax highlight.